pyproject.toml
[project]
name = "nuclear-eng-ai-assistant"
version = "0.1.0"
description = "Nuclear Engineering AI Assistant"
requires-python = ">=3.11"
dependencies = [
    "fastapi",
    "uvicorn",
    "pydantic>=2.0",
    "pydantic-settings",
    "google-cloud-aiplatform",
    "google-cloud-firestore",
    "google-cloud-bigquery",
    "google-cloud-storage",
    "google-cloud-documentai",
    "google-adk",
    "google-genai",
    "pdfplumber",
    "PyPDF2"
]

[project.optional-dependencies]
dev = [
    "pytest",
    "ruff",
    "black",
    "mypy"
]

[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"
